feat(navigation): add navigation package and adopt app-owned NavigationStacks with typed routers
- Add GitLabNavigation module and create routers for Explore, Home, and Profile with typed destinations and a simple navigate API
- Move NavigationStacks to the app target and register all destinations for Explore, Home, and Profile
- Update Explore: use ExploreEntry and push project details from ExploreProjectsView through the router. Remove feature-owned coordinator and stack
- Update Home: add HomeEntry and keep HomeRootView stack-free. Make ProjectsListView accept a navigationContext and route to details through the router
- Update Profile: add ProfileEntry and make ProfileRootView stack-free. Wire settings and personal projects to the router
- Keep features UI-only. Composition and destination mapping live in the app.
