+++ GitLabMobile/Core/GitLabNavigation/Tests/GitLabNavigationTests/GitLabNavigationTests.swift
@@ -1,0 +1,6 @@
+import Testing
+@testable import GitLabNavigation
+
+@Test func example() async throws {
+    // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+}
