+++ GitLabMobile/Features/AuthFeature/Sources/AuthFeature/Localization/AuthStrings.swift
@@ -1,3 +1,11 @@
+//
+//  AuthStrings.swift
+//  AuthFeature
+//
+//  Copyright © 2025 Eliomane. All rights reserved.
+//  Licensed under Apache License v2.0. See LICENSE file.
+//
+
 import Foundation
 
 extension LocalizedStringResource {
