+++ GitLabMobile/Features/ExploreFeature/Sources/ExploreFeature/Presentation/Views/ExploreRootView.swift
@@ -8,20 +8,30 @@
 
 import SwiftUI
 import GitLabDesignSystem
-import ProjectDetailsFeature
+import GitLabNavigation
 import ProjectsDomain
 
 public struct ExploreRootView: View {
     @Environment(ProjectsDependencies.self) private var projectsDependencies
-    @State private var coordinator = ExploreCoordinator()
+    @Environment(ExploreRouter.self) private var router
 
     public init() {}
 
     public var body: some View {
-        NavigationStack(path: Bindable(coordinator).navigationPath) {
-            List {
-                Section {
-                    ForEach(ExploreCoordinator.Entry.allCases, id: \.self) { entry in
+        List {
+            Section {
+                ForEach(ExploreEntry.allCases, id: \.self) { entry in
+                    switch entry {
+                    case .projects:
+                        NavigationLink(value: ExploreRouter.Destination.projects) {
+                            NavigationRow(
+                                systemImage: entry.systemImage,
+                                iconColor: entry.iconColor,
+                                title: entry.title,
+                                subtitle: entry.subtitle
+                            )
+                        }
+                    default:
                         NavigationLink(value: entry.destination) {
                             NavigationRow(
                                 systemImage: entry.systemImage,
@@ -32,26 +42,10 @@ public struct ExploreRootView: View {
                         }
                     }
                 }
-                .listSectionSeparator(.hidden, edges: .top)
             }
-            .navigationBarTitleDisplayMode(.large)
-            .navigationTitle(String(localized: .ExploreL10n.title))
-            .navigationDestination(for: ExploreCoordinator.Destination.self) { destination in
-                switch destination {
-                case .projects:
-                    ExploreProjectsView(repository: projectsDependencies.repository)
-                case .projectDetail(let project):
-                    ProjectDetailsView(projectId: project.id, repository: projectsDependencies.repository)
-                case .users:
-                    Text(.ExploreDestinationsL10n.users)
-                case .groups:
-                    Text(.ExploreDestinationsL10n.groups)
-                case .topics:
-                    Text(.ExploreDestinationsL10n.topics)
-                case .snippets:
-                    Text(.ExploreDestinationsL10n.snippets)
-                }
-            }
+            .listSectionSeparator(.hidden, edges: .top)
         }
+        .navigationBarTitleDisplayMode(.large)
+        .navigationTitle(String(localized: .ExploreL10n.title))
     }
 }
