+++ GitLabMobile/Features/HomeFeature/Sources/HomeFeature/Localization/HomeStrings.swift
@@ -1,3 +1,11 @@
+//
+//  HomeStrings.swift
+//  HomeFeature
+//
+//  Copyright © 2025 Eliomane. All rights reserved.
+//  Licensed under Apache License v2.0. See LICENSE file.
+//
+
 import Foundation
 
 extension LocalizedStringResource {
