+++ GitLabMobile/Features/ProfileFeature/Sources/ProfileFeature/Localization/ProfileStrings.swift
@@ -1,3 +1,11 @@
+//
+//  ProfileStrings.swift
+//  ProfileFeature
+//
+//  Copyright © 2025 Eliomane. All rights reserved.
+//  Licensed under Apache License v2.0. See LICENSE file.
+//
+
 import Foundation
 
 extension LocalizedStringResource {
