+++ GitLabMobile/Features/ProjectDetailsFeature/Sources/ProjectDetailsFeature/Presentation/Stores/ProjectDetailsStore.swift
@@ -36,5 +36,3 @@ public final class ProjectDetailsStore {
         }
     }
 }
-
-
