+++ GitLabMobile/Features/UserProjectsFeature/Sources/UserProjectsFeature/Localization/UserProjectsStrings.swift
@@ -1,3 +1,11 @@
+//
+//  UserProjectsStrings.swift
+//  UserProjectsFeature
+//
+//  Copyright © 2025 Eliomane. All rights reserved.
+//  Licensed under Apache License v2.0. See LICENSE file.
+//
+
 import Foundation
 
 extension LocalizedStringResource {
