+++ GitLabMobile/GitLabMobile/ContentView.swift
@@ -12,19 +12,13 @@ import ExploreFeature
 import ProfileFeature
 
 struct ContentView: View {
+
     var body: some View {
         TabView {
-            HomeRootView()
-                .tabItem { Label(String(localized: .AppTabsL10n.home), systemImage: "house") }
-
-            NavigationStack { Text(String(localized: .AppTabsL10n.notifications)) }
-                .tabItem { Label(String(localized: .AppTabsL10n.notifications), systemImage: "bell") }
-
-            ExploreRootView()
-                .tabItem { Label(String(localized: .AppTabsL10n.explore), systemImage: "binoculars") }
-
-            ProfileRootView()
-                .tabItem { Label(String(localized: .AppTabsL10n.profile), systemImage: "person") }
+            ForEach(AppTab.allCases) { tab in
+                tab.destination()
+                    .tabItem { tab.label }
+            }
         }
     }
 }
