+++ GitLabMobile/GitLabMobile/Localization/AppStrings.swift
@@ -1,3 +1,11 @@
+//
+//  AppStrings.swift
+//  GitLabMobile
+//
+//  Copyright © 2025 Eliomane. All rights reserved.
+//  Licensed under Apache License v2.0. See LICENSE file.
+//
+
 import Foundation
 
 extension LocalizedStringResource {
